fix: 주제가 생성될 때마다 전체 리스트를 불러오는 현상 수정
